+++ packages/config/src/chains.ts
@@ -165,6 +165,7 @@ export const alphanetRelay = new Parachain({
   ecosystem: Ecosystem.AlphanetRelay,
   genesisHash:
     '0xe1ea3ab1d46ba8f4898b6b4b9c54ffc05282d299f89e84bd0fd08067758c9443',
+  isRelay: true,
   isTestChain: true,
   key: 'alphanet-relay',
   name: 'Alphanet Relay',
@@ -749,6 +750,7 @@ export const kusama = new Parachain({
   ecosystem: Ecosystem.Kusama,
   genesisHash:
     '0xb0a8d493285c2df73290dfb7e61f870f17b41801197a149ca93654499ea3dafe',
+  isRelay: true,
   key: 'kusama',
   name: 'Kusama',
   nativeAsset: ksm,
@@ -781,6 +783,7 @@ export const kusamaAssetHub = new Parachain({
       decimals: 12,
     }),
   ],
+  checkSovereignAccountBalances: true,
   ecosystem: Ecosystem.Kusama,
   genesisHash:
     '0x48239ef607d7928874027a43a67689209727dfb3d3dc5e5b03a39bdc2eda771a',
@@ -1963,6 +1966,7 @@ export const polkadot = new Parachain({
   ecosystem: Ecosystem.Polkadot,
   genesisHash:
     '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3',
+  isRelay: true,
   key: 'polkadot',
   name: 'Polkadot',
   nativeAsset: dot,
@@ -2030,6 +2034,7 @@ export const polkadotAssetHub = new Parachain({
       },
     }),
   ],
+  checkSovereignAccountBalances: true,
   ecosystem: Ecosystem.Polkadot,
   genesisHash:
     '0x68d56f15f85d3136970ec16946040bc1752654e906147f7e43e9d539d7c3de2f',