+++ packages/sdk/src/getTransferData/getDestinationData.ts
@@ -1,4 +1,4 @@
-import type { AssetRoute } from '@moonbeam-network/xcm-config';
+import type { AssetRoute, DestinationConfig, SourceConfig } from '@moonbeam-network/xcm-config';
 import type { DestinationChainTransferData } from '../sdk.interfaces';
 import {
   getAssetMin,
@@ -6,6 +6,10 @@ import {
   getDestinationFee,
   getExistentialDeposit,
 } from './getTransferData.utils';
+import {
+  getSovereignAccountAddresses,
+} from '@moonbeam-network/xcm-utils';
+import { Parachain } from '@moonbeam-network/xcm-types';
 
 export interface GetDestinationDataParams {
   route: AssetRoute;
@@ -44,5 +48,56 @@ export async function getDestinationData({
     existentialDeposit,
     fee,
     min,
+    sovereignAccountBalances:  await getSovereignAccountBalances({
+        source: route.source,
+        destination: route.destination,
+      }),
+  };
+  
+}
+
+interface GetSovereignAccountBalancesProps {
+ source: SourceConfig;
+ destination: DestinationConfig;
+}
+
+async function getSovereignAccountBalances( {
+  destination,
+  source,
+}: GetSovereignAccountBalancesProps) {
+
+  if(!Parachain.is(source.chain) || !Parachain.is(destination.chain)) {
+    return undefined
+  }
+
+  const sovereignAccountAddresses = getSovereignAccountAddresses(
+    source.chain.parachainId,
+  );
+
+  const destinationFeeAssetBalance =
+    destination.fee.balance;
+
+  const sovereignAccountAddress = destination.chain.isRelay
+    ? sovereignAccountAddresses.relay
+    : sovereignAccountAddresses.generic;
+
+  const sovereignAccountBalance = await getBalance({
+    address: sovereignAccountAddress,
+    asset: destination.chain.getChainAsset(destination.asset),
+    builder: destination.balance,
+    chain: destination.chain,
+  });
+
+  const sovereignAccountFeeAssetBalance = destinationFeeAssetBalance
+    ? await getBalance({
+        address: sovereignAccountAddress,
+        asset: destination.chain.getChainAsset(destination.fee.asset),
+        builder: destinationFeeAssetBalance,
+    chain: destination.chain,
+      })
+    : undefined;
+  return {
+    feeAssetBalance: sovereignAccountFeeAssetBalance?.amount,
+    transferAssetBalance: sovereignAccountBalance.amount,
   };
 }