+++ packages/sdk/src/getTransferData/getTransferData.ts
@@ -11,7 +11,7 @@ import { EvmService } from '../services/evm/EvmService';
 import { PolkadotService } from '../services/polkadot';
 import { getDestinationData } from './getDestinationData';
 import { getSourceData } from './getSourceData';
-import { convertToChainDecimals, getMin } from './getTransferData.utils';
+import { convertToChainDecimals, getMin, validateSovereignAccountBalances} from './getTransferData.utils';
 
 export interface GetTransferDataParams {
   route: AssetRoute;
@@ -49,6 +49,7 @@ export async function getTransferData({
       const bigAmount = Big(
         toBigInt(amount, sourceData.balance.decimals).toString(),
       );
+
       const result = bigAmount.minus(
         sourceData.balance.isSame(destinationFee)
           ? destinationFee.toBig()
@@ -66,9 +67,15 @@ export async function getTransferData({
       amount,
       { evmSigner, polkadotSigner }: Partial<Signers>,
     ): Promise<string> {
+
       const source = route.source.chain as AnyParachain;
       const destination = route.destination.chain as AnyParachain;
       const bigintAmount = toBigInt(amount, sourceData.balance.decimals);
+      validateSovereignAccountBalances({
+        amount: bigintAmount,
+        destination,
+        source,
+      });
       const asset = AssetAmount.fromChainAsset(
         route.source.chain.getChainAsset(route.source.asset),
         { amount: bigintAmount },